Decouple RemoteStore from implementations
