+++ k9mail/src/main/java/com/fsck/k9/Account.java
@@ -29,6 +29,7 @@ import com.fsck.k9.mail.NetworkType;
 import com.fsck.k9.mail.filter.Base64;
 import com.fsck.k9.mail.ssl.LocalKeyStore;
 import com.fsck.k9.mail.store.RemoteStore;
+import com.fsck.k9.mail.store.RemoteStoreManager;
 import com.fsck.k9.mail.store.StoreConfig;
 import com.fsck.k9.mailstore.LocalStore;
 import com.fsck.k9.mailstore.StorageManager;
@@ -1186,7 +1187,7 @@ public class Account implements BaseAccount, StoreConfig {
     }
 
     public RemoteStore getRemoteStore() throws MessagingException {
-        return RemoteStore.getInstance(K9.app, this);
+        return RemoteStoreManager.getInstance(K9.app, this);
     }
 
     // It'd be great if this actually went into the store implementation
