+++ mail/common/src/test/java/com/fsck/k9/mail/store/imap/ImapStoreUriTest.java
@@ -7,8 +7,6 @@ import java.util.Map;
 import com.fsck.k9.mail.AuthType;
 import com.fsck.k9.mail.ConnectionSecurity;
 import com.fsck.k9.mail.ServerSettings;
-import com.fsck.k9.mail.store.RemoteStore;
-
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
@@ -20,7 +18,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapNoAuth() {
         String uri = "imap://user:pass@server/";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.PLAIN, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -32,7 +30,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapNoPassword() {
         String uri = "imap://user:@server/";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.PLAIN, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -44,7 +42,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapPlainNoPassword() {
         String uri = "imap://PLAIN:user:@server/";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.PLAIN, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -56,7 +54,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapExternalAuth() {
         String uri = "imap://EXTERNAL:user:clientCertAlias@server/";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.EXTERNAL, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -69,7 +67,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapXOAuth2() {
         String uri = "imap://XOAUTH2:user:@server/";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.XOAUTH2, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -81,7 +79,7 @@ public class ImapStoreUriTest {
     @Test
     public void testDecodeStoreUriImapSSL() {
         String uri = "imap+tls+://PLAIN:user:pass@server/";
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(ConnectionSecurity.STARTTLS_REQUIRED, settings.connectionSecurity);
         assertEquals(AuthType.PLAIN, settings.authenticationType);
@@ -94,7 +92,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapTLS() {
         String uri = "imap+ssl+://PLAIN:user:pass@server/";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(ConnectionSecurity.SSL_TLS_REQUIRED, settings.connectionSecurity);
         assertEquals(AuthType.PLAIN, settings.authenticationType);
@@ -107,7 +105,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapAllExtras() {
         String uri = "imap://PLAIN:user:pass@server:143/0%7CcustomPathPrefix";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.PLAIN, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -122,7 +120,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapNoExtras() {
         String uri = "imap://PLAIN:user:pass@server:143/";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.PLAIN, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -136,7 +134,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapPrefixOnly() {
         String uri = "imap://PLAIN:user:pass@server:143/customPathPrefix";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.PLAIN, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -151,7 +149,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapEmptyPrefix() {
         String uri = "imap://PLAIN:user:pass@server:143/0%7C";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.PLAIN, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -166,7 +164,7 @@ public class ImapStoreUriTest {
     public void testDecodeStoreUriImapAutodetectAndPrefix() {
         String uri = "imap://PLAIN:user:pass@server:143/1%7CcustomPathPrefix";
         
-        ServerSettings settings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings settings = ImapStore.decodeUri(uri);
 
         assertEquals(AuthType.PLAIN, settings.authenticationType);
         assertEquals("user", settings.username);
@@ -185,7 +183,7 @@ public class ImapStoreUriTest {
         ServerSettings settings = new ServerSettings(ServerSettings.Type.IMAP, "server", 143,
                 ConnectionSecurity.NONE, AuthType.PLAIN, "user", "pass", null, extra);
 
-        String uri = RemoteStore.createStoreUri(settings);
+        String uri = ImapStore.createUri(settings);
 
         assertEquals("imap://PLAIN:user:pass@server:143/0%7CcustomPathPrefix", uri);
     }
@@ -198,7 +196,7 @@ public class ImapStoreUriTest {
         ServerSettings settings = new ServerSettings(ServerSettings.Type.IMAP, "server", 143,
                 ConnectionSecurity.NONE, AuthType.PLAIN, "user", "pass", null, extra);
 
-        String uri = RemoteStore.createStoreUri(settings);
+        String uri = ImapStore.createUri(settings);
 
         assertEquals("imap://PLAIN:user:pass@server:143/0%7C", uri);
     }
@@ -208,7 +206,7 @@ public class ImapStoreUriTest {
         ServerSettings settings = new ServerSettings(ServerSettings.Type.IMAP, "server", 143,
                 ConnectionSecurity.NONE, AuthType.PLAIN, "user", "pass", null);
 
-        String uri = RemoteStore.createStoreUri(settings);
+        String uri = ImapStore.createUri(settings);
 
         assertEquals("imap://PLAIN:user:pass@server:143/1%7C", uri);
     }
@@ -221,7 +219,7 @@ public class ImapStoreUriTest {
         ServerSettings settings = new ServerSettings(ServerSettings.Type.IMAP, "server", 143,
                 ConnectionSecurity.NONE, AuthType.PLAIN, "user", "pass", null, extra);
 
-        String uri = RemoteStore.createStoreUri(settings);
+        String uri = ImapStore.createUri(settings);
 
         assertEquals("imap://PLAIN:user:pass@server:143/1%7C", uri);
     }
@@ -231,11 +229,11 @@ public class ImapStoreUriTest {
         ServerSettings settings = new ServerSettings(ServerSettings.Type.IMAP, "server", 143,
                 ConnectionSecurity.NONE, AuthType.PLAIN, "user@doma:n", "p@ssw:rd%", null, null);
 
-        String uri = RemoteStore.createStoreUri(settings);
+        String uri = ImapStore.createUri(settings);
 
         assertEquals("imap://PLAIN:user%2540doma%253An:p%2540ssw%253Ard%2525@server:143/1%7C", uri);
 
-        ServerSettings outSettings = RemoteStore.decodeStoreUri(uri);
+        ServerSettings outSettings = ImapStore.decodeUri(uri);
 
         assertEquals("user@doma:n", outSettings.username);
         assertEquals("p@ssw:rd%", outSettings.password);
